Rename docs.py as _icu_docs.py
"docs" is a very generic name for a top-level module.
Rename it as something clearly PyICU-specific and private.
